AI-Creation/requirements.txt
numpy==1.21.0
pandas==1.3.0
scikit-learn==1.0.2
pefile==2021.9.3
python-magic==0.4.24
joblib==1.0.1
matplotlib==3.4.2
seaborn==0.11.1
